Function that checks the dates in the data base and excludes incorrect records, returns the corrected database

Function that checks the dates in the data base and excludes incorrect records, returns the corrected database

checkDates(base, convertDates = FALSE, dateFormat = "",
  bestGuess = FALSE, overnight = FALSE, deleteErrors = "",
  patientID = "pID", hospitalID = "hID", disDate = "Ddate",
  admDate = "Adate", ...)

Arguments

base
(data.table). A patient discharge database, in the form of a data.table. The data.table should have at least the following columns: pID: patientID (character) hID: hospitalID (character) Adate: admission date (POSIXct, but character can be converted to POSIXct) Ddate: discharge date (POSIXct, but character can be converted to POSIXct)

convertDates
boolean indicating if dates need to be converted to POSIXct if they are not

dateFormat
character giving the input format of the date character string

bestGuess
boolean indicating if the function needs to estimate the best date format, based on the best results from 1000 random records.

overnight
boolean indicating if only patient who stayed overnight should be included. TRUE will delete any record with admission and discharge on the same day

deleteErrors
character indicating the way incorrect records should be deleted: "record" deletes just the incorrect record "patient" deletes all records of each patient with one or more incorrect records.

patientID
(charachter) the columns name containing the patient ID. Default is "pID"

hospitalID
(charachter) the columns name containing the hospital ID. Default is "hID"

disDate
(charachter) the columns name containing the discharge date. Default is "Ddate"

admDate
(charachter) the columns name containing the admission date. Default is "Adate"

...
other parameters passed on to internal functions

Value

The corrected database as data.table.