Compute network metrics

Function computing different network analysis metrics.

Usage

get_metrics(
  network,
  mode = "directed",
  weighted = TRUE,
  transfers = TRUE,
  metrics = c("degree", "closeness", "clusters", "betweenness"),
  clusters = c("cluster_fast_greedy", "cluster_infomap"),
  hubs = "all_clusters",
  options = list(degree = list(modes = c("in", "out", "total")), closeness = list(modes =
    "total"), betweenness = list(), cluster_fast_greedy = list(undirected = "collapse"),
    cluster_infomap = list(undirected = "collapse"), clusters = list(algos =
    c("cluster_fast_greedy", "cluster_infomap"), undirected = "collapse"))
)

Arguments

network

the network to analyze. Must be an igraph, HospiNet or a square adjacency matrix (n*n).

mode

either "directed" or "undirected" network measures

weighted

TRUE if the network is weighted

transfers

TRUE if metrics specific to subject transfers must be computed

metrics

list of the metrics to compute

clusters

choose between cluster algorithm: cluster_fast_greedy or cluster_infomap

hubs

choose between getting hubs from "all_clusters" or "global"

options

named list of options to be passed to the igraph functions